refactor(mantine): quality pass - consolidate mocks, extract AnnotationBadges
- Create centralized /mocks/ directory with 10 mock data files
- Extract AnnotationBadges component for reuse in Tools and Resources pages
- Update all pages/components to import from centralized mocks
- Add path alias configuration (@/) to tsconfig.json and vite.config.ts

Files updated:
- New: client/src/mocks/* (servers, tools, resources, logs, prompts, history, tasks, auth, capabilities, index)
- New: client/src/components/AnnotationBadges.tsx
- Updated: All pages (ServerList, Tools, Resources, Logs, Prompts, History, Tasks)
- Updated: Modal components (OAuth, Sampling, Elicitation, Roots, ExperimentalFeatures)
- Updated: tsconfig.json, vite.config.ts
