+++ client/src/components/RootsConfigurationModal.tsx
@@ -18,23 +18,12 @@ import {
   Divider,
   Alert,
 } from '@mantine/core';
-
-interface Root {
-  name: string;
-  uri: string;
-}
+import { initialRoots, type Root } from '@/mocks';
 
 interface RootsConfigurationModalProps {
   open: boolean;
   onOpenChange: (open: boolean) => void;
 }
-
-// Mock initial roots data
-const initialRoots: Root[] = [
-  { name: 'Project', uri: 'file:///home/user/myproject' },
-  { name: 'Documents', uri: 'file:///home/user/Documents' },
-  { name: 'Config', uri: 'file:///etc/myapp' },
-];
 
 export function RootsConfigurationModal({
   open,
