+++ client/src/mocks/index.ts
@@ -1,0 +1,10 @@
+// Re-export all mocks from a single entry point
+export * from './servers';
+export * from './tools';
+export * from './resources';
+export * from './logs';
+export * from './prompts';
+export * from './history';
+export * from './tasks';
+export * from './auth';
+export * from './capabilities';
