+++ client/src/pages/ServerList.tsx
@@ -11,39 +11,7 @@ import { ServerCard, ConnectionMode } from '../components/ServerCard';
 import { AddServerModal, ServerConfig } from '../components/AddServerModal';
 import { ImportServerJsonModal } from '../components/ImportServerJsonModal';
 import { showInfoToast, showSuccessToast } from '../lib/toast';
-
-// Mock server data
-const mockServers = [
-  {
-    id: 'everything-server',
-    name: 'everything-server',
-    version: '1.0.0',
-    transport: 'stdio' as const,
-    command: 'npx -y @modelcontextprotocol/server-everything',
-    status: 'connected' as const,
-    capabilities: { tools: 4, resources: 12, prompts: 2 },
-  },
-  {
-    id: 'filesystem-server',
-    name: 'filesystem-server',
-    version: '0.6.2',
-    transport: 'stdio' as const,
-    command: 'npx -y @modelcontextprotocol/server-filesystem /tmp',
-    status: 'disconnected' as const,
-    capabilities: null,
-  },
-  {
-    id: 'remote-server',
-    name: 'remote-server',
-    version: '2.1.0',
-    transport: 'http' as const,
-    url: 'https://api.example.com/mcp',
-    status: 'failed' as const,
-    retryCount: 3,
-    error: 'Connection timeout after 20s',
-    capabilities: null,
-  },
-];
+import { mockServers } from '@/mocks';
 
 export function ServerList() {
   const [addModalOpen, setAddModalOpen] = useState(false);
